requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Appium-Python-Client==1.1.0
attrs==20.3.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-runner==5.3.0
pytest-variables==1.9.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.4

PyYAML~=5.4.1